+++ module/game_stat.rb
