+++ module/opponent_stat.rb
