+++ test/game_stat_test.rb
@@ -4,7 +4,7 @@ require_relative '../lib/stat_tracker'
 require_relative '../lib/team'
 require_relative '../lib/game'
 require_relative '../lib/game_team'
-require_relative '../lib/game_stat'
+require_relative '../module/game_stat'
 require 'pry'
 
 class GameStatTest < Minitest::Test
