+++ test/opponent_stat_test.rb
@@ -4,6 +4,7 @@ require_relative '../lib/stat_tracker'
 require_relative '../lib/team'
 require_relative '../lib/game'
 require_relative '../lib/game_team'
+require_relative '../module/opponent_stat'
 require 'pry'
 
 class OpponentStatTest < Minitest::Test
