+++ test/team_stat_test.rb
@@ -3,8 +3,7 @@ require 'minitest/pride'
 require_relative '../lib/stat_tracker'
 require_relative '../lib/team'
 require_relative '../lib/game'
-require_relative '../lib/game_team'
-# require_relative '../lib/team_stat'
+require_relative '../module/team_stat'
 require 'pry'
 
 class TeamStatTest < Minitest::Test
